NH-101159: refactor
